requirement.txt
kafka-python==0.9.5
kazoo
